requirements.txt
Django==1.5.5
South==0.8.4
djangorestframework==2.3.10
wsgiref==0.1.2
